feat(components): Refactor Ratio component and remove RatioDisplay
This commit refactors the Ratio component by removing the unused RatioDisplay component. The calculateReadableRatio function is no longer used, and the winningPercentage calculation has been simplified. The text-slate-600 class in the Ratio component now has a text-xs class added to it. The font size of the RatioDisplay component has been changed to text-4xl md:text-6xl, and the tabular-nums class is applied to the percentage value.

In the RatioBar component, the shakeWhenFull variable and its associated class have been removed.

Finally, in the tailwind.config.js file, the "shake-all-directions" animation has been removed.