feat: Add revalidation option to getTweet function
This commit adds a revalidation option to the getTweet function in order to allow for revalidation of tweet data. The revalidation option is passed as an argument to the getTweet function, allowing for more control over the revalidation process. This change ensures that the getTweet function can handle revalidation properly and improves the overall functionality of the QuotedPage component.