feat(components): Add targetPercentage prop to RatioBar
This commit adds a new prop `targetPercentage` to RatioBar component, which represents the target percentage for the width of the winning portion of the ratio bar. It is used to determine the animation behavior of the exploding emoji element. The targetPercentage is compared with the winningPercentage to determine whether the animation should be applied or not.

Additionally, some CSS classes related to the exploding emoji animation have been added to `globals.css`.